% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charcount.R
\name{charcount}
\alias{charcount}
\alias{charcount,partition-method}
\alias{charcount,subcorpus-method}
\alias{charcount,partition_bundle-method}
\alias{charcount,subcorpus_bundle-method}
\alias{charcount,corpus-method}
\title{Count the number of characters}
\usage{
charcount(x, ...)

\S4method{charcount}{partition}(
  x,
  p_attribute = "word",
  char_regex = "[a-zA-Z]",
  lowercase = TRUE,
  decreasing = TRUE
)

\S4method{charcount}{subcorpus}(
  x,
  p_attribute = "word",
  char_regex = "[a-zA-Z]",
  lowercase = TRUE,
  decreasing = TRUE
)

\S4method{charcount}{partition_bundle}(x, mc = FALSE, progress = TRUE, decreasing = TRUE, ...)

\S4method{charcount}{subcorpus_bundle}(x, decreasing = TRUE, mc = FALSE, progress = TRUE, ...)

\S4method{charcount}{corpus}(
  x,
  p_attribute = "word",
  lowercase = TRUE,
  char_regex = "[a-zA-Z]",
  decreasing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{Object to process.}

\item{...}{Argument passed into \code{blapply()}.}

\item{p_attribute}{the p-attribute}

\item{char_regex}{if NULL, counts for all charactrs will be returned, else a regex indicating which characters to include in the counting}

\item{lowercase}{whether to lower tokens}

\item{decreasing}{logical, passed into order call}

\item{mc}{logical}

\item{progress}{A \code{logical} value.}

\item{verbose}{Whether to output progress messages.}
}
\description{
Count the number of characters
}
\examples{
library(polmineR)
use("RcppCWB")

partition("REUTERS", id = "127") \%>\%
  charcount()
  
corpus("REUTERS") \%>\%
  subset(id == "127") \%>\%
  charcount()
  
corpus("REUTERS") \%>\%
  partition_bundle(s_attribute = "id") \%>\%
  charcount()
  
corpus("REUTERS") \%>\%
  split(s_attribute = "id") \%>\%
  charcount()
library(polmineR)
use("RcppCWB")
n <- corpus("REUTERS") \%>\% charcount(decreasing = FALSE)
}
