% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charfilter}
\alias{charfilter}
\title{Reduce characters in vector to a given set of characters}
\usage{
charfilter(x, chars)
}
\arguments{
\item{x}{A \code{character} vector.}

\item{chars}{Vector with characters to keep.}
}
\description{
Reduce characters in vector to a given set of characters
}
\examples{
library(polmineR)
use("duplicates") # make REUTERS2 available

vocab <- corpus("REUTERS2") |>
  p_attributes(p_attribute = "word") |> 
  charfilter(chars = tolower(LETTERS[1:10]))
}
