% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{duplicates_as_annotation_data}
\alias{duplicates_as_annotation_data}
\title{Make annotation data}
\usage{
duplicates_as_annotation_data(x, corpus, s_attribute)
}
\arguments{
\item{x}{Input \code{data.table}.}

\item{corpus}{ID of CWB corpus.}

\item{s_attribute}{Structural attribute to annotate.}
}
\description{
Turn \code{data.table} with duplicates into file with corpus positions and
annotation of duplicates.
}
\examples{
library(polmineR)
use(pkg = "duplicates")

charcount <- corpus("REUTERS2") \%>\% 
  charcount(
    p_attribute = "word",
    char_regex = "[a-zA-Z]",
    lowercase = TRUE,
    decreasing = FALSE
   )

vocab <- corpus("REUTERS2") \%>\% 
  p_attributes(p_attribute = "word") \%>\% 
  charfilter(chars = names(charcount[1:12]))

x <- corpus("REUTERS2") |>
  split(s_attribute = "doc_id")

dupl <- detect_duplicates(
    x = x,
    p_attribute = "word",
    s_attribute = "doc_id",
    mc = parallel::detectCores() - 2L,
    vocab = vocab
  )

grps <- as_docgroups(dupl)

annodata <- duplicates_as_annotation_data(
  x = grps,
  corpus = "REUTERS2", 
  s_attribute = "doc_id"
)
}
