% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{duplicates_encode}
\alias{duplicates_encode}
\title{Encode annotation data}
\usage{
duplicates_encode(x, corpus, method = "R", ...)
}
\arguments{
\item{x}{Data.}

\item{corpus}{ID of CWB corpus.}

\item{method}{XXX.}
}
\description{
Add structural attributes to CWB corpus based on the annotation data that
has been generated.
}
