% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duplicates_get_groups}
\alias{duplicates_get_groups}
\title{Get duplicate groups}
\usage{
duplicates_get_groups(x)
}
\arguments{
\item{x}{A \code{data.table} with duplicates that have been detected.}
}
\description{
Get duplicate groups
}
