\name{LynchSmithEldridgeTavareFragDup-package}
\alias{duplicates-package}
\alias{duplicates}
\docType{package}
\title{
Package to create Sweave for manuscript 
}
\description{
See Sweave for details}
\details{
\tabular{ll}{
Package: \tab duplicates\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-05\cr
License: \tab What license is it under?\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
