% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minimize_vocabulary}
\alias{minimize_vocabulary}
\title{Reduce vocabulary to a set of characters}
\usage{
minimize_vocabulary(x, chars, p_attribute = "word")
}
\arguments{
\item{x}{A CWB corpus.}

\item{chars}{Vector with characters to keep.}

\item{p_attribute}{A positional attribute.}
}
\description{
Reduce vocabulary to a set of characters
}
