\name{secondbiggest}
\alias{secondbiggest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Returns the second largest value from a vector
}
\description{
Returns the second largest value from a vector
}
\usage{
secondbiggest(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a vector of numeric values
}
}
\details{
See Sweave for details
}
\value{
The second largest value in the vector
}
\references{
See the sweave in this package as well as Lynch, Smith et al.
}
\author{
Andy Lynch
}
\note{
no notes
}
\examples{
## see vignette
}

