% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicates.R
\name{as_docgroups}
\alias{as_docgroups}
\title{Get groups of near-duplicate documents}
\usage{
as_docgroups(x)
}
\arguments{
\item{x}{A \code{data.table} with duplicates that have been detected.}
}
\description{
Get groups of near-duplicate documents
}
