% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duplicates_as_annotation_data}
\alias{duplicates_as_annotation_data}
\title{Make annotation data}
\usage{
duplicates_as_annotation_data(
  x,
  corpus,
  s_attribute,
  drop = NULL,
  cols = c("size", "name"),
  order = c(1L, 1L)
)
}
\arguments{
\item{x}{Input \code{data.table}.}

\item{corpus}{ID of CWB corpus.}

\item{s_attribute}{Structural attribute to annotate.}

\item{drop}{A character vector of document IDs that will be removed from
the annotation data. Useful for removing known noise that will be
excluded from the analysis otherwise.}

\item{cols}{XXX.}

\item{order}{XXX.}
}
\description{
Turn \code{data.table} with duplicates into file with corpus positions and
annotation of duplicates.
}
